Feature/client css20

+++ client/src/pages/videopage2.js
@@ -7,10 +7,13 @@ import Comments from '../component/comments'
 axios.defaults.withCredentials = true;
 
 export default function VideoPage2({clickMyVideoData, userInfo, accessToken, viewStateFunc, isLogin, videoInfo}){
-
+//   console.log(clickMyVideoData)
+  
     useEffect(() => {
         viewStateFunc(clickMyVideoData.id);
       }, []) 
+
+      
       
     const history = useHistory();
 
@@ -19,7 +22,7 @@ export default function VideoPage2({clickMyVideoData, userInfo, accessToken, vie
     const deleteMyVideoFunc = () => {
         console.log(accessToken)
         console.log(clickMyVideoData)
-        axios.delete(`https://localhost:80/myvideo?id=${clickMyVideoData.id}`,{
+        axios.delete(`${process.env.REACT_APP_SERVER}/myvideo?id=${clickMyVideoData.id}`,{
             headers: {
             authorization: `Bearer ${accessToken}`,
             "Content-Type" : "application/json"   
@@ -30,7 +33,7 @@ export default function VideoPage2({clickMyVideoData, userInfo, accessToken, vie
           .then((res) => {
               console.log(res)
               if(res.data.message === "deletion is complete") {
-                window.location.replace('/') 
+                window.location.replace('/main') 
               }
           })
           .catch((err) => {
@@ -40,7 +43,7 @@ export default function VideoPage2({clickMyVideoData, userInfo, accessToken, vie
 
     const [likeVideo, setLikeVideo] = useState(false)
     const [likeVideoCount, setLikeVideoCount] = useState(clickMyVideoData.likes)
-
+    // console.log(likeVideo)
    // 좋아요 이모티콘  상태변경 함수
     const likeStateFunc = () => {
         setLikeVideo(!likeVideo);
@@ -57,7 +60,8 @@ export default function VideoPage2({clickMyVideoData, userInfo, accessToken, vie
     // likeVideo가 false일 때 동영상 좋아요 post 요청
     const likeVideoFunc = () => {
         console.log(likeVideo)
-        if(likeVideo === false) {
+        console.log(clickMyVideoData.mychoice)
+        if(likeVideo === false && clickMyVideoData.mychoice === false) {
             axios.post("https://localhost:80/like/video",{video_id: clickMyVideoData.id},
             {
               headers: {
@@ -67,11 +71,10 @@ export default function VideoPage2({clickMyVideoData, userInfo, accessToken, vie
               withCredentials: true
             })
             .then((res) => {
-                console.log(res)
                 likeStateFunc(); // 불이 켜지고
                 likeCountPlusFunc(); // 좋아요 카운트 +1
             })
-        } else if(likeVideo === true) {
+        } else if(likeVideo === false && clickMyVideoData.mychoice === true) {
             axios.delete(`https://localhost:80/like/video?video_id=${clickMyVideoData.id}`,
             {
               headers: {
@@ -81,9 +84,23 @@ export default function VideoPage2({clickMyVideoData, userInfo, accessToken, vie
               withCredentials: true
             })
             .then((res) => {
-                console.log(res)
+                // console.log(res)
                 likeStateFunc(); // 불이 꺼지고
                 likeCountMinusFunc(); // 좋아요 카운트 -1
+            })
+        }
+        else if(likeVideo === true && clickMyVideoData.mychoice === true) {
+            axios.post("https://localhost:80/like/video",{video_id: clickMyVideoData.id},
+            {
+              headers: {
+              authorization: `Bearer ${accessToken}`,
+              "Content-Type" : "application/json"   
+              },
+              withCredentials: true
+            })
+            .then((res) => {
+                likeStateFunc(); // 불이 켜지고
+                likeCountPlusFunc(); // 좋아요 카운트 +1
             })
         }
     }
@@ -108,9 +125,9 @@ export default function VideoPage2({clickMyVideoData, userInfo, accessToken, vie
                     <div className='video2-second-row-creator'><img className="user-icon" src="https://i.ibb.co/ZV9MknX/profile-user.png"/> {userInfo.name}</div>
                     <div className="video2-second-row-like-delete">
                       <div className="like-myvideo btn2" onClick={likeVideoFunc}>
-                      {likeVideo === false ? 
-                      <img className="like-icon" src="https://i.ibb.co/C0ntKBK/2021-11-11-5-56-57-removebg-preview.png" alt="" /> :
-                      <img className="like-icon" src="https://i.ibb.co/Y0jmDXG/2021-11-11-5-58-41-removebg-preview.png" alt="" />
+                      {clickMyVideoData.mychoice === true && likeVideo === false ? 
+                      <img className="like-icon" src="https://i.ibb.co/Y0jmDXG/2021-11-11-5-58-41-removebg-preview.png" alt="" /> :
+                      <img className="like-icon" src="https://i.ibb.co/C0ntKBK/2021-11-11-5-56-57-removebg-preview.png" alt="" />
                       }
                       {likeVideoCount}
                       </div>
@@ -121,7 +138,7 @@ export default function VideoPage2({clickMyVideoData, userInfo, accessToken, vie
                     </div>
                 </div>
             </div>
-            <Comments accessToken={accessToken} videoInfo={videoInfo}/>
+            <Comments accessToken={accessToken} videoInfo={videoInfo} userInfo={userInfo}/>
         </div>
     )
 }
